Conflicts:
	README.md
	src/mongo/s/mongos_main.cpp
	src/third_party/SConscript

+++ jstests/core/columnstore_eligibility.js
@@ -7,15 +7,23 @@
  *   # Cannot run aggregate with explain in a transaction.
  *   does_not_support_transactions,
  *   # column store indexes are still under a feature flag and require full sbe
- *   uses_column_store_index,
  *   featureFlagColumnstoreIndexes,
  *   featureFlagSbeFull,
+ *   # Columnstore tests set server parameters to disable columnstore query planning heuristics -
+ *   # server parameters are stored in-memory only so are not transferred onto the recipient.
+ *   tenant_migration_incompatible,
+ *   not_allowed_with_security_token,
  * ]
  */
 (function() {
 "use strict";
 
 load("jstests/libs/analyze_plan.js");
+load("jstests/libs/columnstore_util.js");  // For setUpServerForColumnStoreIndexTest.
+
+if (!setUpServerForColumnStoreIndexTest(db)) {
+    return;
+}
 
 const coll = db.columnstore_eligibility;
 coll.drop();
@@ -160,4 +168,10 @@ assert(!planHasStage(db, explain, "COLUMN_SCAN"), explain);
 // SBE is not supported for the count command.
 explain = coll.explain().find({a: 2}).count();
 assert(!planHasStage(db, explain, "COLUMN_SCAN"), explain);
+
+// Count-like queries that can use another index should prefer that to column scan.
+assert.commandWorked(coll.createIndex({a: 1}));
+explain = coll.explain().aggregate([{$match: {a: 1}}, {$count: "count"}]);
+assert(aggPlanHasStage(explain, "COUNT_SCAN") || aggPlanHasStage(explain, "IXSCAN"), explain);
+assert.commandWorked(coll.dropIndex({a: 1}));
 }());
